Move viewV2 tests out of row.spec.ts and into viewV2.spec.ts.
