+++ packages/server/src/api/routes/tests/row.spec.ts
@@ -392,6 +392,23 @@ describe.each([
         expect(row.arrayFieldArrayStrKnown).toEqual(["One"])
         expect(row.optsFieldStrKnown).toEqual("Alpha")
       })
+
+    isInternal &&
+      it("doesn't allow creating in user table", async () => {
+        const userTableId = InternalTable.USER_METADATA
+        const response = await config.api.row.save(
+          userTableId,
+          {
+            tableId: userTableId,
+            firstName: "Joe",
+            lastName: "Joe",
+            email: "joe@joe.com",
+            roles: {},
+          },
+          { status: 400 }
+        )
+        expect(response.message).toBe("Cannot create new user entry.")
+      })
   })
 
   describe("get", () => {
@@ -887,642 +904,6 @@ describe.each([
         { rows: [existing._id!] },
         { status: 404 }
       )
-    })
-  })
-
-  describe("view 2.0", () => {
-    async function userTable(): Promise<Table> {
-      return saveTableRequest({
-        name: `users_${uuid.v4()}`,
-        type: "table",
-        schema: {
-          name: {
-            type: FieldType.STRING,
-            name: "name",
-          },
-          surname: {
-            type: FieldType.STRING,
-            name: "surname",
-          },
-          age: {
-            type: FieldType.NUMBER,
-            name: "age",
-          },
-          address: {
-            type: FieldType.STRING,
-            name: "address",
-          },
-          jobTitle: {
-            type: FieldType.STRING,
-            name: "jobTitle",
-          },
-        },
-      })
-    }
-
-    const randomRowData = () => ({
-      name: generator.first(),
-      surname: generator.last(),
-      age: generator.age(),
-      address: generator.address(),
-      jobTitle: generator.word(),
-    })
-
-    describe("create", () => {
-      it("should persist a new row with only the provided view fields", async () => {
-        const table = await config.api.table.save(await userTable())
-        const view = await config.api.viewV2.create({
-          tableId: table._id!,
-          name: generator.guid(),
-          schema: {
-            name: { visible: true },
-            surname: { visible: true },
-            address: { visible: true },
-          },
-        })
-
-        const data = randomRowData()
-        const newRow = await config.api.row.save(view.id, {
-          tableId: table!._id,
-          _viewId: view.id,
-          ...data,
-        })
-
-        const row = await config.api.row.get(table._id!, newRow._id!)
-        expect(row).toEqual({
-          name: data.name,
-          surname: data.surname,
-          address: data.address,
-          tableId: table!._id,
-          _id: newRow._id,
-          _rev: newRow._rev,
-          id: newRow.id,
-          ...defaultRowFields,
-        })
-        expect(row._viewId).toBeUndefined()
-        expect(row.age).toBeUndefined()
-        expect(row.jobTitle).toBeUndefined()
-      })
-    })
-
-    describe("patch", () => {
-      it("should update only the view fields for a row", async () => {
-        const table = await config.api.table.save(await userTable())
-        const tableId = table._id!
-        const view = await config.api.viewV2.create({
-          tableId: tableId,
-          name: generator.guid(),
-          schema: {
-            name: { visible: true },
-            address: { visible: true },
-          },
-        })
-
-        const newRow = await config.api.row.save(view.id, {
-          tableId,
-          _viewId: view.id,
-          ...randomRowData(),
-        })
-        const newData = randomRowData()
-        await config.api.row.patch(view.id, {
-          tableId,
-          _viewId: view.id,
-          _id: newRow._id!,
-          _rev: newRow._rev!,
-          ...newData,
-        })
-
-        const row = await config.api.row.get(tableId, newRow._id!)
-        expect(row).toEqual({
-          ...newRow,
-          name: newData.name,
-          address: newData.address,
-          _id: newRow._id,
-          _rev: expect.any(String),
-          id: newRow.id,
-          ...defaultRowFields,
-        })
-        expect(row._viewId).toBeUndefined()
-        expect(row.age).toBeUndefined()
-        expect(row.jobTitle).toBeUndefined()
-      })
-    })
-
-    describe("destroy", () => {
-      it("should be able to delete a row", async () => {
-        const table = await config.api.table.save(await userTable())
-        const tableId = table._id!
-        const view = await config.api.viewV2.create({
-          tableId: tableId,
-          name: generator.guid(),
-          schema: {
-            name: { visible: true },
-            address: { visible: true },
-          },
-        })
-
-        const createdRow = await config.api.row.save(table._id!, {})
-        const rowUsage = await getRowUsage()
-
-        await config.api.row.bulkDelete(view.id, { rows: [createdRow] })
-
-        await assertRowUsage(rowUsage - 1)
-
-        await config.api.row.get(tableId, createdRow._id!, {
-          status: 404,
-        })
-      })
-
-      it("should be able to delete multiple rows", async () => {
-        const table = await config.api.table.save(await userTable())
-        const tableId = table._id!
-        const view = await config.api.viewV2.create({
-          tableId: tableId,
-          name: generator.guid(),
-          schema: {
-            name: { visible: true },
-            address: { visible: true },
-          },
-        })
-
-        const rows = await Promise.all([
-          config.api.row.save(table._id!, {}),
-          config.api.row.save(table._id!, {}),
-          config.api.row.save(table._id!, {}),
-        ])
-        const rowUsage = await getRowUsage()
-
-        await config.api.row.bulkDelete(view.id, { rows: [rows[0], rows[2]] })
-
-        await assertRowUsage(rowUsage - 2)
-
-        await config.api.row.get(tableId, rows[0]._id!, {
-          status: 404,
-        })
-        await config.api.row.get(tableId, rows[2]._id!, {
-          status: 404,
-        })
-        await config.api.row.get(tableId, rows[1]._id!, { status: 200 })
-      })
-    })
-
-    describe("view search", () => {
-      let table: Table
-      const viewSchema = { age: { visible: true }, name: { visible: true } }
-
-      beforeAll(async () => {
-        table = await config.api.table.save(
-          saveTableRequest({
-            name: `users_${uuid.v4()}`,
-            schema: {
-              name: {
-                type: FieldType.STRING,
-                name: "name",
-                constraints: { type: "string" },
-              },
-              age: {
-                type: FieldType.NUMBER,
-                name: "age",
-                constraints: {},
-              },
-            },
-          })
-        )
-      })
-
-      it("returns empty rows from view when no schema is passed", async () => {
-        const rows = await Promise.all(
-          Array.from({ length: 10 }, () =>
-            config.api.row.save(table._id!, { tableId: table._id })
-          )
-        )
-
-        const createViewResponse = await config.api.viewV2.create({
-          tableId: table._id!,
-          name: generator.guid(),
-        })
-        const response = await config.api.viewV2.search(createViewResponse.id)
-
-        expect(response.rows).toHaveLength(10)
-        expect(response).toEqual({
-          rows: expect.arrayContaining(
-            rows.map(r => ({
-              _viewId: createViewResponse.id,
-              tableId: table._id,
-              _id: r._id,
-              _rev: r._rev,
-              ...defaultRowFields,
-            }))
-          ),
-          ...(isInternal
-            ? {}
-            : {
-                hasNextPage: false,
-                bookmark: null,
-              }),
-        })
-      })
-
-      it("searching respects the view filters", async () => {
-        await Promise.all(
-          Array.from({ length: 10 }, () =>
-            config.api.row.save(table._id!, {
-              tableId: table._id,
-              name: generator.name(),
-              age: generator.integer({ min: 10, max: 30 }),
-            })
-          )
-        )
-
-        const expectedRows = await Promise.all(
-          Array.from({ length: 5 }, () =>
-            config.api.row.save(table._id!, {
-              tableId: table._id,
-              name: generator.name(),
-              age: 40,
-            })
-          )
-        )
-
-        const createViewResponse = await config.api.viewV2.create({
-          tableId: table._id!,
-          name: generator.guid(),
-          query: [
-            { operator: SearchQueryOperators.EQUAL, field: "age", value: 40 },
-          ],
-          schema: viewSchema,
-        })
-
-        const response = await config.api.viewV2.search(createViewResponse.id)
-
-        expect(response.rows).toHaveLength(5)
-        expect(response).toEqual({
-          rows: expect.arrayContaining(
-            expectedRows.map(r => ({
-              _viewId: createViewResponse.id,
-              tableId: table._id,
-              name: r.name,
-              age: r.age,
-              _id: r._id,
-              _rev: r._rev,
-              ...defaultRowFields,
-            }))
-          ),
-          ...(isInternal
-            ? {}
-            : {
-                hasNextPage: false,
-                bookmark: null,
-              }),
-        })
-      })
-
-      const sortTestOptions: [
-        {
-          field: string
-          order?: SortOrder
-          type?: SortType
-        },
-        string[]
-      ][] = [
-        [
-          {
-            field: "name",
-            order: SortOrder.ASCENDING,
-            type: SortType.STRING,
-          },
-          ["Alice", "Bob", "Charly", "Danny"],
-        ],
-        [
-          {
-            field: "name",
-          },
-          ["Alice", "Bob", "Charly", "Danny"],
-        ],
-        [
-          {
-            field: "name",
-            order: SortOrder.DESCENDING,
-          },
-          ["Danny", "Charly", "Bob", "Alice"],
-        ],
-        [
-          {
-            field: "name",
-            order: SortOrder.DESCENDING,
-            type: SortType.STRING,
-          },
-          ["Danny", "Charly", "Bob", "Alice"],
-        ],
-        [
-          {
-            field: "age",
-            order: SortOrder.ASCENDING,
-            type: SortType.number,
-          },
-          ["Danny", "Alice", "Charly", "Bob"],
-        ],
-        [
-          {
-            field: "age",
-            order: SortOrder.ASCENDING,
-          },
-          ["Danny", "Alice", "Charly", "Bob"],
-        ],
-        [
-          {
-            field: "age",
-            order: SortOrder.DESCENDING,
-          },
-          ["Bob", "Charly", "Alice", "Danny"],
-        ],
-        [
-          {
-            field: "age",
-            order: SortOrder.DESCENDING,
-            type: SortType.number,
-          },
-          ["Bob", "Charly", "Alice", "Danny"],
-        ],
-      ]
-
-      describe("sorting", () => {
-        let table: Table
-        beforeAll(async () => {
-          table = await config.api.table.save(await userTable())
-          const users = [
-            { name: "Alice", age: 25 },
-            { name: "Bob", age: 30 },
-            { name: "Charly", age: 27 },
-            { name: "Danny", age: 15 },
-          ]
-          await Promise.all(
-            users.map(u =>
-              config.api.row.save(table._id!, {
-                tableId: table._id,
-                ...u,
-              })
-            )
-          )
-        })
-
-        it.each(sortTestOptions)(
-          "allow sorting (%s)",
-          async (sortParams, expected) => {
-            const createViewResponse = await config.api.viewV2.create({
-              tableId: table._id!,
-              name: generator.guid(),
-              sort: sortParams,
-              schema: viewSchema,
-            })
-
-            const response = await config.api.viewV2.search(
-              createViewResponse.id
-            )
-
-            expect(response.rows).toHaveLength(4)
-            expect(response.rows).toEqual(
-              expected.map(name => expect.objectContaining({ name }))
-            )
-          }
-        )
-
-        it.each(sortTestOptions)(
-          "allow override the default view sorting (%s)",
-          async (sortParams, expected) => {
-            const createViewResponse = await config.api.viewV2.create({
-              tableId: table._id!,
-              name: generator.guid(),
-              sort: {
-                field: "name",
-                order: SortOrder.ASCENDING,
-                type: SortType.STRING,
-              },
-              schema: viewSchema,
-            })
-
-            const response = await config.api.viewV2.search(
-              createViewResponse.id,
-              {
-                sort: sortParams.field,
-                sortOrder: sortParams.order,
-                sortType: sortParams.type,
-                query: {},
-              }
-            )
-
-            expect(response.rows).toHaveLength(4)
-            expect(response.rows).toEqual(
-              expected.map(name => expect.objectContaining({ name }))
-            )
-          }
-        )
-      })
-
-      it("when schema is defined, defined columns and row attributes are returned", async () => {
-        const table = await config.api.table.save(await userTable())
-        const rows = await Promise.all(
-          Array.from({ length: 10 }, () =>
-            config.api.row.save(table._id!, {
-              tableId: table._id,
-              name: generator.name(),
-              age: generator.age(),
-            })
-          )
-        )
-
-        const view = await config.api.viewV2.create({
-          tableId: table._id!,
-          name: generator.guid(),
-          schema: { name: { visible: true } },
-        })
-        const response = await config.api.viewV2.search(view.id)
-
-        expect(response.rows).toHaveLength(10)
-        expect(response.rows).toEqual(
-          expect.arrayContaining(
-            rows.map(r => ({
-              ...(isInternal
-                ? expectAnyInternalColsAttributes
-                : expectAnyExternalColsAttributes),
-              _viewId: view.id,
-              name: r.name,
-            }))
-          )
-        )
-      })
-
-      it("views without data can be returned", async () => {
-        const table = await config.api.table.save(await userTable())
-        const createViewResponse = await config.api.viewV2.create({
-          tableId: table._id!,
-          name: generator.guid(),
-        })
-        const response = await config.api.viewV2.search(createViewResponse.id)
-        expect(response.rows).toHaveLength(0)
-      })
-
-      it("respects the limit parameter", async () => {
-        const table = await config.api.table.save(await userTable())
-        await Promise.all(
-          Array.from({ length: 10 }, () => config.api.row.save(table._id!, {}))
-        )
-
-        const limit = generator.integer({ min: 1, max: 8 })
-
-        const createViewResponse = await config.api.viewV2.create({
-          tableId: table._id!,
-          name: generator.guid(),
-        })
-        const response = await config.api.viewV2.search(createViewResponse.id, {
-          limit,
-          query: {},
-        })
-
-        expect(response.rows).toHaveLength(limit)
-      })
-
-      it("can handle pagination", async () => {
-        const table = await config.api.table.save(await userTable())
-        await Promise.all(
-          Array.from({ length: 10 }, () => config.api.row.save(table._id!, {}))
-        )
-        const view = await config.api.viewV2.create({
-          tableId: table._id!,
-          name: generator.guid(),
-        })
-        const rows = (await config.api.viewV2.search(view.id)).rows
-
-        const page1 = await config.api.viewV2.search(view.id, {
-          paginate: true,
-          limit: 4,
-          query: {},
-        })
-        expect(page1).toEqual({
-          rows: expect.arrayContaining(rows.slice(0, 4)),
-          totalRows: isInternal ? 10 : undefined,
-          hasNextPage: true,
-          bookmark: expect.anything(),
-        })
-
-        const page2 = await config.api.viewV2.search(view.id, {
-          paginate: true,
-          limit: 4,
-          bookmark: page1.bookmark,
-
-          query: {},
-        })
-        expect(page2).toEqual({
-          rows: expect.arrayContaining(rows.slice(4, 8)),
-          totalRows: isInternal ? 10 : undefined,
-          hasNextPage: true,
-          bookmark: expect.anything(),
-        })
-
-        const page3 = await config.api.viewV2.search(view.id, {
-          paginate: true,
-          limit: 4,
-          bookmark: page2.bookmark,
-          query: {},
-        })
-        expect(page3).toEqual({
-          rows: expect.arrayContaining(rows.slice(8)),
-          totalRows: isInternal ? 10 : undefined,
-          hasNextPage: false,
-          bookmark: expect.anything(),
-        })
-      })
-
-      isInternal &&
-        it("doesn't allow creating in user table", async () => {
-          const userTableId = InternalTable.USER_METADATA
-          const response = await config.api.row.save(
-            userTableId,
-            {
-              tableId: userTableId,
-              firstName: "Joe",
-              lastName: "Joe",
-              email: "joe@joe.com",
-              roles: {},
-            },
-            { status: 400 }
-          )
-          expect(response.message).toBe("Cannot create new user entry.")
-        })
-
-      describe("permissions", () => {
-        let table: Table
-        let view: ViewV2
-
-        beforeAll(async () => {
-          table = await config.api.table.save(await userTable())
-          await Promise.all(
-            Array.from({ length: 10 }, () =>
-              config.api.row.save(table._id!, {})
-            )
-          )
-
-          view = await config.api.viewV2.create({
-            tableId: table._id!,
-            name: generator.guid(),
-          })
-        })
-
-        beforeEach(() => {
-          mocks.licenses.useViewPermissions()
-        })
-
-        it("does not allow public users to fetch by default", async () => {
-          await config.publish()
-          await config.api.viewV2.publicSearch(view.id, undefined, {
-            status: 403,
-          })
-        })
-
-        it("allow public users to fetch when permissions are explicit", async () => {
-          await config.api.permission.add({
-            roleId: roles.BUILTIN_ROLE_IDS.PUBLIC,
-            level: PermissionLevel.READ,
-            resourceId: view.id,
-          })
-          await config.publish()
-
-          const response = await config.api.viewV2.publicSearch(view.id)
-
-          expect(response.rows).toHaveLength(10)
-        })
-
-        it("allow public users to fetch when permissions are inherited", async () => {
-          await config.api.permission.add({
-            roleId: roles.BUILTIN_ROLE_IDS.PUBLIC,
-            level: PermissionLevel.READ,
-            resourceId: table._id!,
-          })
-          await config.publish()
-
-          const response = await config.api.viewV2.publicSearch(view.id)
-
-          expect(response.rows).toHaveLength(10)
-        })
-
-        it("respects inherited permissions, not allowing not public views from public tables", async () => {
-          await config.api.permission.add({
-            roleId: roles.BUILTIN_ROLE_IDS.PUBLIC,
-            level: PermissionLevel.READ,
-            resourceId: table._id!,
-          })
-          await config.api.permission.add({
-            roleId: roles.BUILTIN_ROLE_IDS.POWER,
-            level: PermissionLevel.READ,
-            resourceId: view.id,
-          })
-          await config.publish()
-
-          await config.api.viewV2.publicSearch(view.id, undefined, {
-            status: 403,
-          })
-        })
-      })
     })
   })
